App restructure.
