+++ client-src/src/app/app.component.ts
@@ -7,10 +7,10 @@ import { InqProvider } from '../providers/inq/inq';
 import { FollowUpProvider } from '../providers/follow-up/follow-up';
 
 import { HomePage } from '../pages/home/home';
-import { InqForm1Page } from '../pages/inq-form1/inq-form1';
-import { InqDetailsPage } from '../pages/inq-details/inq-details';
-import { InqListPage } from '../pages/inq-list/inq-list';
-import { InqSummaryPage } from '../pages/inq-summary/inq-summary';
+import { InqForm1Page } from '../pages/visitor/inq-form1/inq-form1';
+import { InqDetailsPage } from '../pages/counselor/inq-details/inq-details';
+import { InqListPage } from '../pages/counselor/inq-list/inq-list';
+import { InqSummaryPage } from '../pages/counselor/inq-summary/inq-summary';
 
 @Component({
   templateUrl: 'app.html'
