+++ client-src/src/app/app.module.ts
@@ -6,14 +6,14 @@ import { Ng2CompleterModule } from "ng2-completer";
 
 import { MyApp } from './app.component';
 import { HomePage } from '../pages/home/home';
-import { InqForm1Page } from '../pages/inq-form1/inq-form1';
-import { InqForm2Page } from '../pages/inq-form2/inq-form2';
-import { InqForm3Page } from '../pages/inq-form3/inq-form3';
-import { InqDetailsPage } from '../pages/inq-details/inq-details';
-import { InqListPage } from '../pages/inq-list/inq-list';
-import { InqSummaryPage } from '../pages/inq-summary/inq-summary';
-import { ThankyouPage } from '../pages/thankyou/thankyou';
-import { InqCloseModalPage } from '../pages/inq-close-modal/inq-close-modal';
+import { InqForm1Page } from '../pages/visitor/inq-form1/inq-form1';
+import { InqForm2Page } from '../pages/visitor/inq-form2/inq-form2';
+import { InqForm3Page } from '../pages/visitor/inq-form3/inq-form3';
+import { InqDetailsPage } from '../pages/counselor/inq-details/inq-details';
+import { InqListPage } from '../pages/counselor/inq-list/inq-list';
+import { InqSummaryPage } from '../pages/counselor/inq-summary/inq-summary';
+import { ThankyouPage } from '../pages/visitor/thankyou/thankyou';
+import { InqCloseModalPage } from '../pages/counselor/inq-close-modal/inq-close-modal';
 
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
