+++ client-src/src/pages/counselor/inq-close-modal/inq-close-modal.scss
