+++ client-src/src/pages/counselor/inq-details/inq-details.scss
