+++ client-src/src/pages/counselor/inq-summary/inq-summary.scss
