+++ client-src/src/pages/visitor/thankyou/thankyou.scss
