+++ client-src/src/pages/visitor/thankyou/thankyou.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
 
-import { HomePage } from '../../pages/home/home';
-import { InqForm1Page } from '../../pages/inq-form1/inq-form1';
+// import { HomePage } from '../../pages/home/home';
+import { InqForm1Page } from '../inq-form1/inq-form1';
 
 @Component({
   selector: 'page-thankyou',
@@ -17,9 +17,9 @@ export class ThankyouPage {
     console.log('ionViewDidLoad ThankyouPage');
   }
 
-  goToHome() {
-    this.navCtrl.setRoot(HomePage);
-  }
+  // goToHome() {
+  //   this.navCtrl.setRoot(HomePage);
+  // }
   newInquiry() {
     this.navCtrl.setRoot(InqForm1Page);
   }
